// components/NoteForm/NoteForm.tsx
import { useId } from 'react';
import css from './NoteForm.module.css';
import { Formik, Form, Field, ErrorMessage } from 'formik';
import * as Yup from 'yup';
import type { NoteInput } from '../../types/note';
import { useMutation, useQueryClient } from '@tanstack/react-query';
import { createNote } from '../../lib/api';
import toast from 'react-hot-toast';

interface NoteFormProps {
  cancel: () => void;
}

const initValues: NoteInput = {
  title: '',
  content: '',
  tag: 'Todo',
};

const NoteSchema = Yup.object().shape({
  title: Yup.string()
    .min(3, 'Title must be at least 3 characters')
    .max(50, 'Title is too long')
    .required('Title is required'),
  content: Yup.string().max(500, 'Content is too long'),
  tag: Yup.string().oneOf(
    ['Todo', 'Work', 'Personal', 'Meeting', 'Shopping'],
    'Invalid tag'
  ),
});

export default function NoteForm({ cancel }: NoteFormProps) {
  const fieldId = useId();

  const queryClient = useQueryClient();
  const { mutate, isPending } = useMutation({
    mutationFn: (newNote: NoteInput) => createNote(newNote),
    onSuccess: () => {
      queryClient.invalidateQueries({ queryKey: ['note'] });
      cancel();
    },
    onError: () => {
      toast.error('The Note didn`t create.. Try again!');
    },
  });

  const handleCancel = () => {
    cancel();
  };

  const noteFormSubmit = (values: NoteInput) => {
    mutate(values);
  };

  return (
    <>
      <Formik
        initialValues={initValues}
        onSubmit={noteFormSubmit}
        validationSchema={NoteSchema}
      >
        <Form className={css.form}>
          <div className={css.formGroup}>
            <label htmlFor={`${fieldId}-title`}>Title</label>
            <Field
              id={`${fieldId}-title`}
              type="text"
              name="title"
              className={css.input}
            />
            <ErrorMessage component="span" name="title" className={css.error} />
          </div>

          <div className={css.formGroup}>
            <label htmlFor={`${fieldId}-content`}>Content</label>
            <Field
              as="textarea"
              id={`${fieldId}-content`}
              name="content"
              rows={8}
              className={css.textarea}
            />
            <ErrorMessage
              component="span"
              name="content"
              className={css.error}
            />
          </div>

          <div className={css.formGroup}>
            <label htmlFor={`${fieldId}-tag`}>Tag</label>
            <Field
              as="select"
              id={`${fieldId}-tag`}
              name="tag"
              className={css.select}
            >
              <option value="Todo">Todo</option>
              <option value="Work">Work</option>
              <option value="Personal">Personal</option>
              <option value="Meeting">Meeting</option>
              <option value="Shopping">Shopping</option>
            </Field>
            <ErrorMessage component="span" name="tag" className={css.error} />
          </div>

          <div className={css.actions}>
            <button
              type="button"
              className={css.cancelButton}
              onClick={handleCancel}
            >
              Cancel
            </button>
            <button
              type="submit"
              className={css.submitButton}
              disabled={isPending}
            >
              {isPending ? 'Creating new note...' : 'Create note'}
            </button>
          </div>
        </Form>
      </Formik>
    </>
  );
}
